requirements.txt
argon2-cffi==23.1.0
awscli==1.44.59
blinker==1.9.0
boto3==1.42.69
botocore==1.42.69
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
docutils==0.19
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
pyasn1==0.6.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.6
